+++ src/plugins/dialogs.ts
@@ -40,9 +40,9 @@ export class Dialogs {
 
   /**
    * Shows a custom alert or dialog box.
-   * @param message Dialog message. (String)
-   * @param title Dialog title. (String) (Optional, defaults to Alert)
-   * @param buttonName Button name. (String) (Optional, defaults to OK)
+   * @param {string} message Dialog message.
+   * @param {string} title Dialog title. (Optional, defaults to Alert)
+   * @param {string} buttonName Button name. (Optional, defaults to OK)
    * @returns {Promise<any>} Returns a blank promise once the user has dismissed the alert.
    */
   @Cordova({
@@ -57,9 +57,9 @@ export class Dialogs {
 
   /**
    * Displays a customizable confirmation dialog box.
-   * @param message Dialog message. (String)
-   * @param title Dialog title. (String) (Optional, defaults to Confirm)
-   * @param buttonLabels Array of strings specifying button labels. (Array) (Optional, defaults to [OK,Cancel])
+   * @param {string} message Dialog message.
+   * @param {string} title Dialog title. (Optional, defaults to Confirm)
+   * @param {Array<string>} buttonLabels Array of strings specifying button labels. (Optional, defaults to [OK,Cancel])
    * @returns {Promise<number>} Returns a promise that resolves the button index that was clicked. Note that the index use one-based indexing.
    */
   @Cordova({
@@ -74,10 +74,10 @@ export class Dialogs {
 
   /**
    * Displays a native dialog box that is more customizable than the browser's prompt function.
-   * @param message Dialog message. (String)
-   * @param title Dialog title (String) (Optional, defaults to Prompt)
-   * @param buttonLabels  Array of strings specifying button labels (Array) (Optional, defaults to ["OK","Cancel"])
-   * @param defaultText Default textbox input value (String) (Optional, Default: empty string)
+   * @param {string} message Dialog message.
+   * @param {string} title Dialog title. (Optional, defaults to Prompt)
+   * @param {Array<string>} buttonLabels  Array of strings specifying button labels. (Optional, defaults to ["OK","Cancel"])
+   * @param {string} defaultText Default textbox input value.  (Optional, Default: empty string)
    * @returns {Promise<any>} Returns a promise that resolves an object with the button index clicked and the text entered
    */
   @Cordova({
@@ -94,7 +94,7 @@ export class Dialogs {
 
   /**
    * The device plays a beep sound.
-   * @param times The number of times to repeat the beep. (Number)
+   * @param {numbers} times The number of times to repeat the beep.
    */
   @Cordova({
     sync: true