+++ src/plugins/push.ts
@@ -191,24 +191,24 @@ export interface IOSPushOptions {
    * The data required in order to enable Action Buttons for iOS.
    * Action Buttons on iOS - https://github.com/phonegap/phonegap-plugin-push/blob/master/docs/PAYLOAD.md#action-buttons-1
    */
-  categories?: CategoryArray
+  categories?: CategoryArray;
 }
 
 export interface CategoryArray {
-  [name: string]: CategoryAction
+  [name: string]: CategoryAction;
 }
 
 export interface CategoryAction {
-  yes?: CategoryActionData
-  no?: CategoryActionData
-  maybe?: CategoryActionData
+  yes?: CategoryActionData;
+  no?: CategoryActionData;
+  maybe?: CategoryActionData;
 }
 
 export interface CategoryActionData {
-  callback: string
-  title: string
-  foreground: boolean
-  destructive: boolean
+  callback: string;
+  title: string;
+  foreground: boolean;
+  destructive: boolean;
 }
 
 export interface AndroidPushOptions {